+++ src/main/java/com/betbull/market/service/ContractService.java
@@ -5,6 +5,7 @@ import com.betbull.market.model.Contract;
 import com.betbull.market.model.Player;
 import com.betbull.market.model.Team;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -30,7 +31,7 @@ public interface ContractService {
      * @param team   the team
      * @return the double
      */
-    double calculateContractFee(Player player, Team team);
+    BigDecimal calculateContractFee(Player player, Team team);
 
     /**
      * Gets contracts by player id.
